Better ZYActivity.
The activityViewController method in UIActivity is called more than
once by the UIActivityViewController, no it's not stortying the
UIViewController being returned by the programmer.

In order to save the user the task to retain and keep a pointer to the
activityViewController he was returning, now the ZYActivity does that
on its own.